requirements.txt
flask
flask-mysqldb
flask-bcrypt
pyjwt
flask_sqlalchemy
mysql-connector-python
pandas
scikit-learn

